requirements.txt
# ─── Core ───────────────────────────────────────────────────────────────────
python-dotenv==1.0.0

# ─── Data ───────────────────────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
datasets==2.19.1
huggingface-hub==0.23.2
tabulate==0.9.0
langdetect==1.0.9

# ─── NLP & Embeddings ───────────────────────────────────────────────────────
transformers==4.41.2
accelerate==0.31.0
sentence-transformers==3.0.0
tokenizers==0.19.1
# NOTE: Install torch with CUDA separately:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124

# ─── RAG Pipeline ───────────────────────────────────────────────────────────
faiss-cpu==1.8.0

# ─── Classification ─────────────────────────────────────────────────────────
scikit-learn==1.4.2

# ─── Evaluation Metrics ─────────────────────────────────────────────────────
nltk==3.8.1
rouge-score==0.1.2
evaluate==0.4.2

# ─── API ────────────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.3
pydantic-settings==2.2.1

# ─── MLOps ──────────────────────────────────────────────────────────────────
mlflow==2.13.2

# ─── Dashboard ──────────────────────────────────────────────────────────────
streamlit==1.35.0
plotly==5.22.0

# ─── EDA & Visualization ────────────────────────────────────────────────────
matplotlib==3.9.0
seaborn==0.13.2
wordcloud==1.9.3

# ─── Notebook ───────────────────────────────────────────────────────────────
jupyter==1.0.0
ipykernel==6.29.4
tqdm==4.66.4

# ─── Stability ──────────────────────────────────────────────────────────────
threadpoolctl==3.6.0